Add function movement() which makes clock function correctly according to current time
Function movement() fetches the current hours, minutes, and seconds, and maps that to the correct degrees
for each hand. Example: The clock is 360 deg, so for 60 seconds in a minute that is 6 deg per second. 6 deg
is then multiplied by the output from getSeconds() to give what degrees the second hand currently should be.