Add code to make clock "animation" seem smoother mimicking a mechanical style movement
Interval and steps for updating hand postions are now less, and more constant. This makes the movement of the
clock appear smoother as the hands move less each time, but more frequently, like that of a mechanical
movement watch.
User will be able to switch between "mechanical" and "normal" modes in future.